Update to decouple DimensionControl component from knowledge of Block attribute schema
Previously the `DimensionControl` component manually constructed the attribute to be updated via string concatenation. This was brittle and prone to error and also couple the component to use in a Block context.

Updated to utilise partial application to apply the dimension attribute argument in advance for each component. This allows us to simply the API of the component.