Update main project README
Update main project repository reamde with various updates that bring it
current for now and down the line.

First organized it into a more general getting started recognizing the
three levels that people might be coming to the project: as a User, as a
Devleoper, and as a Contributor

For User - added links to the user documentation, support forums, and
how to report an issue

For Develoepr - added links to the developer documentation

For Contributor - simplifed the contributor sections and links to the
contributors guide

I removed the links at the bottom since those are mostly dated from when
the project was starting out and justifying itself. These links can be
included elsewhere for background and context, and not necessary on the
welcoming page.