Try single icon for all sites in the inserter
One of the issues with social links is a duplication of the icons, one
copy for displaying within the content and another copy for displaying
in the inserter.

This commit uses a single icon to see what it looks like in the inserter
if all the sites use the same external share icon.